% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.names.wl.R
\name{convert.names.wl}
\alias{convert.names.wl}
\title{Turns wavelengths into variable's names}
\usage{
convert.names.wl(start = NULL, stop = NULL, step = 2)
}
\arguments{
\item{start}{First wavelength of the spectra.}

\item{stop}{Last wavelength of the spectra.}

\item{step}{Distance between each recorded wavelength.}
}
\value{
Returns vector with syntactically valid names for each wavelength
}
\description{
Turns wavelengths into variable's names
}
\examples{
data(beer)
X=beer[,2:ncol(beer)]
head(names(X))
names(X)=convert.names.wl(1100,2250,2)
head(names(X))
}
